build: use custom build_py command to bundle prebuilt C extension
Unlike `ml_dtypes`, `zk_dtypes` depends on the `abseil-cpp` library and
potentially more libraries. Listing all dependency sources in `setup.py`
is laborious.

Instead, we use a **custom `build_py` command** to bundle the prebuilt
C extension. The build process isn't clean, but it works, and we plan to
improve it later.