build: update setuptools to >= 77.0.3 and adopt PEP 639 support
Updates the required `setuptools` version to 77.0.3 or higher to adopt
the standard defined in PEP 639.

Previously, the `license` metadata either used a deprecated table format
(with 'file' or 'text' keys) or required the full license text. Since
setuptools 77.0.3 introduced support for PEP 639, we can now use the
**SPDX license expression** format for the `license` field, standardizing
the build metadata.