refactor: remove `ZK_DTYPES_USE_ABSL` macro
Removes the `ZK_DTYPES_USE_ABSL` macro, which was originally introduced to
prevent the ZKIR project from introducing an explicit dependency on
`abseil-cpp`.

After recent change(#17) allowing ZKIR to depend on the `all_types` target
(which transitively pulls in `abseil-cpp`), this macro is no longer
necessary and has been removed for cleanup.